pro/requirements.txt
PyYAML==6.0.2
chardet==5.2.0
cryptography==46.0.3
matplotlib-inline==0.1.7
matplotlib==3.10.7
numpy==2.1.3
pandas==2.3.3
pdfplumber==0.11.7
pillow==10.4.0
pyinstaller-hooks-contrib==2025.9
pyinstaller==6.16.0
pyqt6-qt6==6.10.0
pyqt6-sip==13.10.2
pyqt6==6.10.0
pyqtgraph==0.13.3
python-dateutil==2.9.0.post0
reportlab==4.2.2
requests==2.32.3
sqlalchemy==2.0.44